fix: correct minor code quality issues across codebase                                                                                                                                                                                                                                                                     - Fix mock_keys fixture return type hint to type[MockKeys]                                                                                                                                                                                                        - Fix _gravity type annotation from int to float in PlayerModel   - Guard player property against empty sprite group (returns None)                                                                                                                                                                                                 - Rename default logger name from "tkinter-app" to "pygame-app"                                                                                                                                                                                                   - Narrow exception catch from Exception to AttributeError in resource_path                                                                                                                                                                                        - Remove irrelevant migrations exclude from pre-commit ruff hooks

+++ src/models/player_model.py
@@ -18,7 +18,7 @@ class PlayerModel(pygame.sprite.Sprite):
         super().__init__()
 
         self._walk_index: float = 0.0
-        self._gravity: int = 0
+        self._gravity: float = 0.0
 
         self._walk_frames: list[pygame.Surface] = [
             pygame.image.load(GRAPHIC_PLAYER_WALK_1).convert_alpha(),
